+++ api/AltV.Net.Client/Elements/Factories/IPlayerFactory.cs
@@ -4,6 +4,6 @@ namespace AltV.Net.Client.Elements.Factories
 {
     public interface IPlayerFactory : IEntityFactory<IPlayer>
     {
-        public ILocalPlayer GetLocalPlayer(ICore core, IntPtr localPlayerPointer, ushort id);
+        public ILocalPlayer GetLocalPlayer(ICore core, IntPtr localPlayerPointer, uint id);
     }
 }
