+++ api/AltV.Net.Shared/IReadOnlyBaseBaseObjectPool.cs
@@ -5,7 +5,7 @@ namespace AltV.Net.Shared
 {
     public interface IReadOnlyBaseBaseObjectPool
     {
-        ISharedBaseObject GetOrCreate(ISharedCore core, IntPtr entityPointer, BaseObjectType baseObjectType, ushort entityId);
+        ISharedBaseObject GetOrCreate(ISharedCore core, IntPtr entityPointer, BaseObjectType baseObjectType, uint entityId);
         ISharedBaseObject GetOrCreate(ISharedCore core, IntPtr entityPointer, BaseObjectType baseObjectType);
         ISharedBaseObject Get(IntPtr entityPointer, BaseObjectType baseObjectType);
     }
